Hack: Adjust API to work with `github.com/lib/pq`
We had to change internals to use a transaction and advisory locks. This commit
is currently in a state were I have hacked at the API and tests until they
pass. No gaurentees yet that this will work.

Changes:
* Use transactions and `pg_try_advisory_xact_lock`. Required since we can't get
  connections with default sql interface.
* Remove `Done()`. Given we are using transactions, if you call `Delete()` or
  `Error()` we commit at that point.
* Remove `Conn()`. you can no longer bundle along other changes
  inside of the job transaction. We can add this back in, but explicitly
  disallow people committing it. We will also have to do something clever to
  rollback, but allow us recording the error.
* Others that I can't remember

Either way, expect this commit to disappear as I split this out into changes
that actually make sense + give me confidence this isn't just a hack.